+++ src/components/calculation_funcs.c
@@ -1,4 +1,11 @@
 #include "../s21_string.h"
+
+size_t s21_strlen(const char *str) {
+  size_t len = 0;
+  while (str[len] != '\0') len++;
+  return len;
+}
+
 /**
  * @brief Calculates the length of the initial segment of str1 which consists
  * entirely of characters not in str2.
