--- src/components/create.sh
@@ -1,38 +1,0 @@
-# touch s21_strstr.c
-# touch s21_strchr.c
-# touch s21_strrchr.c
-# touch s21_strlen.c
-# touch s21_strpbrk.c
-# touch s21_strcspn.c
-# touch s21_strncpy.c
-# touch s21_strncmp.c
-# touch s21_strtok.c
-# touch s21_memchr.c
-# touch s21_memcmp.c
-# touch s21_memcpy.c
-# touch s21_memset.c
-# touch s21_strncat.c
-# touch s21_to_lower.c
-# touch s21_to_upper.c
-# touch s21_trim.c
-# touch s21_insert.c
-
-
-# s21_strstr.c
-# s21_strchr.c
-# s21_strrchr.c
-# s21_strlen.c
-# s21_strpbrk.c
-# s21_strcspn.c
-# s21_strncpy.c
-# s21_strncmp.c
-# s21_strtok.c
-# s21_memchr.c
-# s21_memcmp.c
-# s21_memcpy.c
-# s21_memset.c
-# s21_strncat.c
-# s21_to_lower.c
-# s21_to_upper.c
-# s21_trim.c
-# s21_insert.c
