--- src/components/s21_strlen.c
@@ -1,7 +1,0 @@
-#include "../s21_string.h"
-
-size_t s21_strlen(const char *str) {
-  size_t len = 0;
-  while (str[len] != '\0') len++;
-  return len;
-}
